Add pickle support and update documentation
This commit integrates the pickle module into the main.py file in order to load the pre-trained model files. This was implemented to remove the need for retraining models during every execution, which significantly improves predict time and reduces resource usage.
Moreover, the 'test' and 'evaluate' methods were modified to use the pickled model, facilitating a more concise workflow.

In addition to code changes, the documentation for the report (td1.md) has been updated. The accuracy results were rephrased to place more emphasis on the effects of preprocessing on the model performance. These changes were made to present a more comprehensible picture of the steps followed and their resultant outcomes in the preliminary part of TD1.